agari-rbac/requirements.txt
Flask==2.3.3
Flask-RESTX==1.3.0
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
